+++ src/front_end/carsdata/src/components/UDScomponents/DivCenterDoors.tsx
@@ -5,13 +5,87 @@ import ModalUDS from './ModalUDS';
 import './style.css';
 import { displayLoadingCircle, displayErrorPopup, removeLoadingCicle } from '../sharedComponents/LoadingCircle';
 
-interface doorsData {
+export interface doorsData {
     ajar: any,
     door: any,
     passenger: any,
     passenger_lock: any,
 }
 
+const checkInput = (message: any) => {
+    let value;
+    do {
+        value = prompt(message);
+        if (value !== '0' && value !== '1') {
+            alert('Accepted value: 0/1');
+        }
+    } while (value !== '0' && value !== '1');
+    return value;
+};
+
+export const writeInfoDoors = async (setData: any) => {
+    displayLoadingCircle();
+    const door = checkInput('Enter Door Parameter:');
+    const passenger = checkInput('Enter Passenger:');
+    const passenger_lock = checkInput('Enter Passenger Lock:');
+    const driver = checkInput('Enter Driver:');
+    const ajar = checkInput('Enter Ajar:');
+    const is_manual_flow = true;
+
+    const data = {
+        door: door || null,
+        passenger: passenger || null,
+        passenger_lock: passenger_lock || null,
+        driver: driver || null,
+        ajar: ajar || null,
+        is_manual_flow: is_manual_flow || null,
+
+    };
+    console.log("Writing info doors...");
+    console.log(data);
+    try {
+        await fetch(`http://127.0.0.1:5000/api/write_info_doors`, {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify(data),
+        }).then(response => response.json())
+            .then(data => {
+                setData(data);
+                console.log(data);
+                // fetchLogs();
+            });
+    } catch (error) {
+        console.log(error);
+        removeLoadingCicle();
+    }
+    removeLoadingCicle();
+
+    readInfoDoors(false, setData)
+}
+
+export const readInfoDoors = async (isManualFlow:boolean, setData:any) => {
+    displayLoadingCircle();
+    console.log("Reading doors info...");
+    try {
+        await fetch(`http://127.0.0.1:5000/api/read_info_doors?is_manual_flow=${isManualFlow}`, {
+            method: 'GET',
+        }).then(response => response.json())
+            .then(data => {
+                console.log(data);
+                setData(data);
+                if (data?.ERROR === "interrupted") {
+                    displayErrorPopup("Connection failed");
+                }
+            });
+    } catch (error) {
+        console.log(error);
+        removeLoadingCicle();
+        displayErrorPopup("Connection failed");
+    }
+    removeLoadingCicle();
+};
 
 const DivCenterDoors = (props: any) => {
     const [data, setData] = useState<doorsData | null>(null);
@@ -19,29 +93,8 @@ const DivCenterDoors = (props: any) => {
     const [error, setError] = useState(null);
 
     useEffect(() => {
-        const readInfoDoors = async () => {
-            displayLoadingCircle();
-            console.log("Reading doors info...");
-            await fetch(`http://127.0.0.1:5000/api/read_info_doors?is_manual_flow=false`)
-                .then(response => {
-                    if (!response.ok) {
-                        throw new Error('Network response was not ok');
-                    }
-                    return response.json();
-                })
-                .then(data => {
-                    setData(data);
-
-                })
-                .catch(error => {
-                    setError(error);
-                    displayErrorPopup("Connection failed");
-                    removeLoadingCicle();
-
-                });
-            removeLoadingCicle();
-        };
-        readInfoDoors();
+        
+        readInfoDoors(false, setData);
     }, []);
